moved utils file to proc module
