+++ src/lib.rs
@@ -39,11 +39,11 @@ mod primitives;
 mod proc;
 mod utils;
 
-pub use self::context::{utils as proc_utils, Context};
+pub use self::context::Context;
 pub use self::errors::Error;
 use self::primitives::Primitive;
 pub use self::sexp::SExp;
-pub use proc::{Arity, Func, Proc};
+pub use proc::{utils as proc_utils, Arity, Func, Proc};
 
 /// A shorthand Result type.
 pub type Result = ::std::result::Result<SExp, Error>;
